[move] constantValue.jsをutilsディレクトリに移動
